+++ src/main/java/gift/controller/gift/GiftController.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package gift.controller.gift;
 
 import gift.dto.paging.PagingRequest;
 import gift.dto.paging.PagingResponse;
